// frontend/src/api-client/types.gen.ts
export type Pagination = {
  count: number;
  currentPage: number;
  nextPage: number;
  pageSize: number;
  pages: number;
  previousPage: number;
};

export type Role = {
  createdAt: string;
  description: string;
  id: string;
  name: string;
  permissions: Array<string>;
  updatedAt: string;
};

export type Roles = Array<Role>;

export type SuccessResponse = {
  item?: User | Role;
  items?: Users | Roles;
  pagination?: Pagination;
};

export type User = {
  bio?: string;
  createdAt: string;
  id: string;
  image?: string;
  mfaEnabled: boolean;
  mfaVerified: boolean;
  name: string;
  updatedAt: string;
  username: string;
};

export type Users = Array<User>;